Completed updated directory structure migration
