+++ src/components/ListView/ListViewModal.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import { Item } from "../../../model/Subject";
+import { Item } from "../../model/Subject";
 import { Modal, getTheme, mergeStyleSets } from "office-ui-fabric-react";
-import SubjectComponent from "../../../components/Subject";
+import SubjectComponent from "../Subject";
 
 const theme = getTheme();
 const styles = mergeStyleSets({
