+++ src/components/ListView/UseListViewRender.tsx
@@ -1,8 +1,8 @@
 import React, { useCallback } from "react";
-import { Item, Subject } from "../../../model/Subject";
+import { Item, Subject } from "../../model/Subject";
 import { IColumn } from "office-ui-fabric-react";
 import ListViewButtons from "./ListViewButtons";
-import { SortField, SortFieldKey } from "../../../model/Order";
+import { SortField, SortFieldKey } from "../../model/Order";
 
 interface UseListView {
   openModal: (item: Item) => void;
