+++ src/components/ListView/UseListViewScroll.ts
@@ -1,8 +1,7 @@
 import { IDetailsList } from "office-ui-fabric-react";
 import { useRef, useEffect } from "react";
-import { getDiffIndex } from "../View";
-import { UseSubjectView } from "../../../components/SubjectView";
-import { SubjectState } from "../../../model/Subject";
+import { UseSubjectView, getDiffIndex } from "../SubjectView";
+import { SubjectState } from "../../model/Subject";
 
 interface UseListViewScroll {
   currentOrder: string[];
