+++ src/components/SimpleListView.tsx
@@ -2,8 +2,8 @@ import React, { useCallback } from "react";
 import { useSelector } from "react-redux";
 import { List, Text } from "office-ui-fabric-react";
 import { mergeStyleSets } from "@uifabric/styling";
-import { State } from "../../Reducer";
-import { getItems, Item } from "../../model/Subject";
+import { State } from "../Reducer";
+import { getItems, Item } from "../model/Subject";
 
 interface ListViewProps {
   parentId?: string;
