+++ src/components/Subject/index.tsx
@@ -15,7 +15,7 @@ import {
   completeWithChildrenItem,
   deleteItem,
   uncompleteItem,
-} from "../../subject/components/SubjectButtonsProps";
+} from "../SubjectButtonsProps";
 import {
   completeSubject,
   uncompleteSubject,
@@ -25,9 +25,9 @@ import { useDispatch, useSelector } from "react-redux";
 
 import { Item } from "../../model/Subject";
 import { Link } from "react-router-dom";
-import SimpleListView from "../../subject/components/SimpleListView";
+import SimpleListView from "../SimpleListView";
 import { State } from "../../Reducer";
-import SubjectListItem from "../../subject/components/ListItem/SubjectListItem";
+import SubjectListItem from "../ListItem/SubjectListItem";
 import TitleInput from "../TitleInput";
 import { deleteSubject } from "../../model/Subject/DeleteSubject";
 import { gotoSubject } from "../../model/Subject/Routing";
