--- src/subject/components/View.ts
@@ -1,14 +1,0 @@
-export const getDiffIndex = (
-  oldOrder: string[],
-  newOrder: string[],
-): number => {
-  for (let i = 0; i < newOrder.length; i++) {
-    const newEl = newOrder[i];
-    const oldEl = oldOrder[i];
-
-    if (newEl !== oldEl) {
-      return i;
-    }
-  }
-  return 0;
-};
