--- src/views/components/ViewsNav/index.ts
@@ -1,3 +1,0 @@
-import ViewsNav from "./ViewsNav";
-
-export default ViewsNav;
